working proc16a pav2. ive refactored existing tests for this. they need to be reviewed and expanded to account for pav2 edge cases

+++ tools/Assembler.Tests/Assembly/CpuRegisterUtilTests.cs
@@ -65,6 +65,6 @@ public class CpuRegisterUtilTests {
 	[Test]
 	[TestCaseSource(nameof(SetBitTestCases))]
 	public void TestSetBit(int input, int bit, bool value, int expectedResult) {
-		Assert.That(AssemblyUtil.SetBit(input, bit, value), Is.EqualTo(expectedResult));
+		Assert.That(AssemblyUtil.SetBit((ushort) input, bit, value), Is.EqualTo(expectedResult));
 	}
 }
